Feature user acl

+++ app/Http/Controllers/AdminCategoryController.php
@@ -21,7 +21,7 @@ class AdminCategoryController extends Controller
             'cat_img' => 'required|image|mimes:jpeg,png,jpg,gif|max:2048',
             'cat_color' => 'required',
         ]);
-        
+
         $name = "";
         if ($request->hasFile('cat_img')) {
             $image = $request->file('cat_img');
@@ -30,7 +30,7 @@ class AdminCategoryController extends Controller
         }else{
             return back()->with('error', 'SORRY - Category Image Required.');
         }
-        
+
 
         categoty::insert([
             'name' => $request->cat_name,
@@ -43,7 +43,7 @@ class AdminCategoryController extends Controller
     }
 
     function Update($id , Request $request){
-    
+
         $request->validate([
             'cat_name' => 'required',
             'cat_color' => 'required',
@@ -67,7 +67,7 @@ class AdminCategoryController extends Controller
             }else{
                 return back()->with('error', 'SORRY - Category Image Required.');
             }
-    
+
             categoty::where('id', $id)->update([
                 'name' => $request->cat_name,
                 'status' => '1',
@@ -83,7 +83,9 @@ class AdminCategoryController extends Controller
     function Drop($id){
         //delete img
         $old_img = categoty::findOrFail($id)->cat_img;
-        unlink('storage/photos/'.$old_img);
+        if (file_exists('storage/photos/'.$old_img)) {
+            unlink('storage/photos/'.$old_img);
+        }
         categoty::findOrFail($id)->delete();
         return back()->with('success', 'Category Deleted Successfully.');
     }
